Filter readout units; reorganize forms
