+++ assays/forms.py
@@ -1,7 +1,12 @@
+import csv
+
 from django import forms
 from django.forms.models import BaseInlineFormSet
 from assays.models import *
 from compounds.models import Compound
+# Use regular expressions for a string split at one point
+import re
+import string
 
 # TODO REFACTOR WHITTLING TO BE HERE IN LIEU OF VIEW
 # TODO REFACTOR FK QUERYSETS TO AVOID N+1
@@ -319,3 +324,403 @@ class PlateTestResultInlineFormset(BaseInlineFormSet):
                 pass
         if results < 1:
             raise forms.ValidationError('You must have at least one result.')
+
+# This function turns a label to a number
+def label_to_number(label):
+    num = 0
+    for char in label:
+        if char in string.ascii_letters:
+            num = num * 26 + (ord(char.upper()) - ord('A')) + 1
+    return num
+
+# TODO CHANGE BLOCK UPLOAD
+# TODO ADD TABULAR UPLOAD
+# TODO LINKING MULTIPLE ASSAYS TO ONE FEATURE IS AMBIGUOUS: DO NOT ALLOW IT (X?)
+# TODO DO NOT ALLOW ROW OR COLUMN OVERFLOW (or underflow?)
+class AssayPlateReadoutInlineFormset(CloneableBaseInlineFormSet):
+    def __init__(self,*args,**kwargs):
+        super (AssayPlateReadoutInlineFormset,self).__init__(*args,**kwargs)
+        unit_queryset = PhysicalUnits.objects.filter(availability__contains='readout').order_by('base_unit', 'scale_factor')
+        for form in self.forms:
+            form.fields['readout_unit'].queryset = unit_queryset
+
+    def clean(self):
+        """Validate unique, existing PLATE READOUTS"""
+
+        # Get upload type
+        upload_type = self.data.get('upload_type')
+
+        forms_data = [f for f in self.forms if f.cleaned_data and not f.cleaned_data.get('DELETE', False)]
+
+        # Dic of assay names from inline with respective unit as value
+        assays = {}
+        # Dic of features with respective assay
+        features_to_assay = {}
+        for form in forms_data:
+            try:
+                if form.cleaned_data:
+                    assay_name = form.cleaned_data.get('assay_id').assay_name
+                    unit = form.cleaned_data.get('readout_unit').unit
+
+                    feature = form.cleaned_data.get('feature')
+
+                    # Error when feature assignation ambiguous
+                    if feature in features_to_assay:
+                        raise forms.ValidationError(
+                            'The feature "{}" is used more than once; ambiguous feature binding is not permitted'.format(feature))
+
+                    features_to_assay.update({feature: assay_name})
+
+                    if assay_name not in assays:
+                        assays.update({assay_name:unit})
+                    else:
+                        raise forms.ValidationError(
+                            'Duplicate assays are not permitted; please blank out or change the duplicate')
+            except AttributeError:
+                pass
+        if len(assays) < 1:
+            raise forms.ValidationError('You must have at least one assay')
+
+        # TODO
+        # If there is already a file in the database and it is not being replaced or cleared (check for clear is implicit)
+        if self.instance.file and not forms_data[-1].files:
+
+            saved_data = AssayReadout.objects.filter(assay_device_readout=self.instance).prefetch_related('assay')
+
+            for raw in saved_data:
+
+                assay = raw.assay.assay_id.assay_name
+                val_unit = raw.assay.readout_unit.unit
+                feature = raw.assay.feature
+
+                # Raise error when an assay does not exist
+                if assay not in assays:
+                    raise forms.ValidationError(
+                        'You can not remove the assay "%s" because it is in your uploaded data.' % assay)
+                # Raise error if val_unit not equal to one listed in APRA
+                if val_unit != assays.get(assay,''):
+                    raise forms.ValidationError(
+                        'The current value unit "%s" does not correspond with the readout unit of "%s"' % (val_unit, assays.get(assay,'')))
+                # TODO Raise error if feature does not correspond?
+                if feature not in features_to_assay:
+                    raise forms.ValidationError(
+                        'You can not remove the feature "{}" because it is in your uploaded data.'.format(feature))
+                if features_to_assay.get(feature) != assay:
+                    raise forms.ValidationError(
+                        'The assay-feature pair in the uploaded data is "{0}-{1}," not "{2}-{3}."'.format(assay,feature,features_to_assay.get(feature),feature))
+
+        # TODO what shall a uniqueness check look like?
+        # If there is a new file
+        if forms_data[-1].files:
+            test_file = forms_data[-1].files.get('file','')
+
+            datareader = csv.reader(test_file, delimiter=',')
+            datalist = list(datareader)
+
+            # Acquire number of rows and columns
+            number_of_rows = self.instance.setup.assay_layout.device.number_of_rows
+            number_of_columns = self.instance.setup.assay_layout.device.number_of_columns
+
+            # Tedious way of getting timeunit; probably should refactor
+            readout_time_unit = PhysicalUnits.objects.get(id=self.data.get('timeunit')).unit
+
+            if upload_type == 'Block':
+                # Number of assays found
+                assays_found = 0
+                # Number of data blocks found
+                data_blocks_found = 0
+
+                for row_index, line in enumerate(datalist):
+                    # If line is blank, skip it
+                    if not line:
+                        continue
+
+                    # If this line is a header
+                    # NOTE THAT FEATURE -> ASSAY
+                    # Headers should look like: FEATURE, {{FEATURE}}, READOUT UNIT, {{READOUT UNIT}}, TIME, {{TIME}}. TIME UNIT, {{TIME UNIT}}
+                    if line[0].lower().strip() == 'feature':
+                        # Throw error if header too short
+                        if len(line) < 4:
+                            raise forms.ValidationError(
+                                'Header row: {} is too short'.format(line))
+
+                        feature = line[1].strip()
+
+                        assay = features_to_assay.get(feature, '')
+                        assays_found += 1
+
+                        val_unit = line[3].strip()
+
+                        # Raise error if feature does not exist
+                        if feature not in features_to_assay:
+                            raise forms.ValidationError(
+                                'No feature with the name "%s" exists; please change your file or add this feature' % feature)
+                        # Raise error when an assay does not exist
+                        if assay not in assays:
+                            raise forms.ValidationError(
+                                'No assay with the name "%s" exists; please change your file or add this assay' % assay)
+                        # Raise error if val_unit not equal to one listed in ACRA
+                        if val_unit != assays.get(assay,''):
+                            raise forms.ValidationError(
+                                'The value unit "%s" does not correspond with the selected readout unit of "%s"' % (val_unit, assays.get(assay,'')))
+
+                        # Fail if time given without time units
+                        if len(line) < 8 and len(line) > 4 and any(line[4:]):
+                            raise forms.ValidationError(
+                                'Header row: {} improperly configured'.format(line))
+
+                        if len(line) >= 8 and any(line[4:]):
+                            time = line[5].strip()
+                            time_unit = line[7].strip()
+
+                            # Fail if time is not numeric
+                            try:
+                                float(time)
+                            except:
+                                raise forms.ValidationError(
+                                    'The time "{}" is invalid. Please only enter numeric times'.format(time))
+
+                            # Fail if time unit does not match
+                            # TODO make a better fuzzy match, right now just checks to see if the first letters correspond
+                            if time_unit[0] != readout_time_unit[0]:
+                                raise forms.ValidationError('The time unit "{}" does not correspond with the selected readout time unit of "{}"'.format(time_unit, readout_time_unit))
+
+                    # Otherwise the line contains datapoints for the current assay
+                    else:
+                        # TODO REVISE HOW DATA_BLOCKS ARE ACQUIRED
+                        if data_blocks_found == 0 or (row_index-assays_found) % number_of_rows == 0:
+                            data_blocks_found += 1
+
+                        # This should handle blocks that have too many rows or do not have a header
+                        if data_blocks_found > assays_found:
+                            raise forms.ValidationError(
+                                        'All plate data must have an assay associated with it. Please add a header line.')
+
+                        # This is to deal with an excess of columns
+                        if len(line) != number_of_columns:
+                            raise forms.ValidationError(
+                                "The number of columns does not correspond with the device's dimensions:{}".format(line))
+
+                        # For every value in the line
+                        for val in line:
+                            # Check every value to make sure it can resolve to a float
+                            try:
+                                # Keep empty strings, though they technically can not be converted to floats
+                                if val != '':
+                                    float(val)
+                            except:
+                                raise forms.ValidationError(
+                                        'The value "%s" is invalid; please make sure all values are numerical' % str(val))
+
+            # If not block, then it is TABULAR data
+            else:
+                # Purge empty lines, they are useless for tabular uploads
+                datalist = [row for row in datalist if any(row)]
+                # The first line SHOULD be the header
+                header = datalist[0]
+
+                if header[1].lower().strip() == 'time':
+                    # The features are the third column of the header onward
+                    features = header[3:]
+                    time_specified = True
+                else:
+                    # The features are the third column of the header onward
+                    features = header[1:]
+                    time_specified = False
+
+                # Exclude the header to get only the data points
+                data = datalist[1:]
+
+                # Fail if there are no features
+                if len(features) < 1:
+                    raise forms.ValidationError(
+                        'The file does not contain any features')
+
+                for feature in features:
+                    # Get associated assay
+                    assay = features_to_assay.get(feature, '')
+                    # Raise error if feature does not exist
+                    if feature not in features_to_assay:
+                        raise forms.ValidationError(
+                            'No feature with the name "%s" exists; please change your file or add this feature' % feature)
+                    # Raise error when an assay does not exist
+                    if assay not in assays:
+                        raise forms.ValidationError(
+                            'No assay with the name "%s" exists; please change your file or add this assay' % assay)
+
+                for row_index, row in enumerate(data):
+                    # Check if well id is valid
+                    try:
+                        # The well identifier given
+                        well = row[0]
+                        # Split the well into alphabetical and numeric
+                        row_label, column_label = re.findall(r"[^\W\d_]+|\d+", well)
+                        # Convert row_label to a number
+                        row_label = label_to_number(row_label)
+                        # Convert column label to an integer
+                        column_label = int(column_label)
+
+                        if row_label > number_of_rows:
+                            raise forms.ValidationError(
+                                "The number of rows does not correspond with the device's dimensions")
+
+                        if column_label > number_of_columns:
+                            raise forms.ValidationError(
+                                "The number of columns does not correspond with the device's dimensions")
+
+                    except:
+                        raise forms.ValidationError(
+                        'Error parsing the well ID: {}'.format(well))
+
+                    if time_specified:
+                        # Values are the slice of the fourth item onward
+                        values = row[3:]
+                        time = row[1]
+                        time_unit = row[2]
+
+                        # Check time unit
+                        # TODO make a better fuzzy match, right now just checks to see if the first letters correspond
+                        if time_unit[0] != readout_time_unit[0]:
+                            raise forms.ValidationError(
+                                'The time unit "%s" does not correspond with the selected readout time unit of "%s"' % (time_unit, readout_time_unit))
+
+                        # Check time
+                        try:
+                            float(time)
+                        except:
+                            raise forms.ValidationError(
+                                'Error while parsing time "{}"'.format(time))
+                    else:
+                        # Values are the slice of the first item onward
+                        values = row[1:]
+                        time = 0
+                        time_unit = None
+
+                    for column_index, value in enumerate(values):
+                        #feature = features[column_index]
+
+                        # Check if all the values can be parsed as floats
+                        try:
+                            if value != '':
+                                float(value)
+                        except:
+                            raise forms.ValidationError(
+                                    'The value {} is invalid; please make sure all values are numerical'.format(value))
+        return self.cleaned_data
+
+
+class AssayChipReadoutInlineFormset(CloneableBaseInlineFormSet):
+    def __init__(self,*args,**kwargs):
+        super (AssayChipReadoutInlineFormset,self).__init__(*args,**kwargs)
+        unit_queryset = PhysicalUnits.objects.filter(availability__contains='readout').order_by('base_unit', 'scale_factor')
+        for form in self.forms:
+            form.fields['readout_unit'].queryset = unit_queryset
+
+    def clean(self):
+        """Validate unique, existing Chip Readout IDs"""
+
+        # Throw error if headers is not valid
+        try:
+            headers = int(self.data.get('headers','')) if self.data.get('headers') else 0
+        except:
+            raise forms.ValidationError('Please make number of headers a valid number.')
+
+        forms_data = [f for f in self.forms if f.cleaned_data and not f.cleaned_data.get('DELETE', False)]
+
+        # Dic of assay names from inline with respective unit as value
+        assays = {}
+        for form in forms_data:
+            try:
+                if form.cleaned_data:
+                    assay_name = form.cleaned_data.get('assay_id').assay_name
+                    unit = form.cleaned_data.get('readout_unit').unit
+                    if assay_name not in assays:
+                        assays.update({assay_name:unit})
+                    else:
+                        raise forms.ValidationError(
+                            'Duplicate assays are not permitted; please blank out or change the duplicate')
+            except AttributeError:
+                pass
+        if len(assays) < 1:
+            raise forms.ValidationError('You must have at least one assay')
+
+        # If there is already a file in the database and it is not being replaced or cleared (check for clear is implicit)
+        if self.instance.file and not forms_data[-1].files:
+            new_time_unit = self.instance.timeunit
+            old_time_unit = AssayChipReadout.objects.get(id=self.instance.id).timeunit
+
+            # Fail if time unit does not match
+            if new_time_unit != old_time_unit:
+                raise forms.ValidationError(
+                    'The time unit "%s" does not correspond with the selected readout time unit of "%s"' % (new_time_unit, old_time_unit))
+
+            saved_data = AssayChipRawData.objects.filter(assay_chip_id=self.instance).prefetch_related('assay_id')
+
+            for raw in saved_data:
+
+                assay = raw.assay_id.assay_id.assay_name
+                val_unit = raw.assay_id.readout_unit.unit
+
+                # Raise error when an assay does not exist
+                if assay not in assays:
+                    raise forms.ValidationError(
+                        'You can not remove the assay "%s" because it is in your uploaded data.' % assay)
+                # Raise error if val_unit not equal to one listed in ACRA
+                if val_unit != assays.get(assay,''):
+                    raise forms.ValidationError(
+                        'The current value unit "%s" does not correspond with the readout unit of "%s"' % (val_unit, assays.get(assay,'')))
+
+        # If there is a new file
+        if forms_data[-1].files:
+            test_file = forms_data[-1].files.get('file','')
+
+            datareader = csv.reader(test_file, delimiter=',')
+            datalist = list(datareader)
+
+            # Tedious way of getting timeunit; probably should refactor
+            readout_time_unit = PhysicalUnits.objects.get(id=self.data.get('timeunit')).unit
+
+            # All unique rows based on ('assay_id', 'field_id', 'elapsed_time')
+            unique = {}
+
+            # Read headers going onward
+            for line in datalist[headers:]:
+
+                # Some lines may not be long enough (have sufficient commas), ignore such lines
+                # Some lines may be empty or incomplete, ignore these as well
+                if len(line) < 6 or not all(line):
+                    continue
+
+                time = line[0]
+                time_unit = line[1].strip().lower()
+                assay = line[2]
+                field = line[3]
+                val = line[4]
+                val_unit = line[5].strip()
+                # Raise error when an assay does not exist
+                if assay not in assays:
+                    raise forms.ValidationError(
+                        'No assay with the name "%s" exists; please change your file or add this assay' % assay)
+                # Raise error if val_unit not equal to one listed in ACRA
+                if val_unit != assays.get(assay,''):
+                    raise forms.ValidationError(
+                        'The value unit "%s" does not correspond with the selected readout unit of "%s"' % (val_unit, assays.get(assay,'')))
+                # Fail if time unit does not match
+                # TODO make a better fuzzy match, right now just checks to see if the first letters correspond
+                if time_unit[0] != readout_time_unit[0]:
+                    raise forms.ValidationError(
+                        'The time unit "%s" does not correspond with the selected readout time unit of "%s"' % (time_unit, readout_time_unit))
+                if (time,assay,field) not in unique:
+                    unique.update({(time,assay,field):True})
+                else:
+                    raise forms.ValidationError(
+                        'File contains duplicate reading %s' % str((time,assay,field)))
+                # Check every value to make sure it can resolve to a float
+                try:
+                    # Keep empty strings, though they technically can not be converted to floats
+                    if val != '':
+                        float(val)
+                except:
+                    raise forms.ValidationError(
+                            'The value "%s" is invalid; please make sure all values are numerical' % str(val))
+        return self.cleaned_data
